# Reduce executable size: compress the included dependency licenses text. 