The current executable size is quite big. One thing which increases the executable size is the DEP_LICENSES file, which is distributed with every build. Some licenses of deendencies we use require this and for others: it is only fair to attribute these authors.  

This file contains huge amounts of duplication (because many dependencies use
the same licenses), thus is likely to have a good compression ratio.

This PR compresses the DEP_LICENSES file on compile time. Then (still on compile time) the resulting compressed bytes are included in the executable. At runtime, if `--dep-licenses` is used, we uncompress the license text and show it. 

If this doesn't reduce the binary enough (but how much is enough?), another option might be to provide the release
binaries within a zip file, which also contains the dependency licenses
text separately.

I currently do not intend to compile for smaller executable sizes over speed optimizations or to go to extremes to reduce the file size.

There are also additional steps, such as stripping symbol information, which
will provide additional gains in reduced file size (stripping release build actually gives huge gains, which is something I should definitely do).

**text only compressed**
```
566K dependency_licenses.txt
89K dependency_licenses.txt.sz
15K dependency_licenses.txt.gz
```

The current implementation uses Snappy compression, since the library was very accessible. However, other compression methods might have even bigger reductions (such as gzip).
Before this PR will be merged, I'd like to try out a few additional compression libraries.

Sizes with `cargo --release`:

|  |  |  | size (bytes) | stripped (bytes) |  |  |
|---------------------|-------------------|---------|----------------|------------------|-------------------|------------------------|
| Encoder crate | Compression level | format | `ls -l` | `ls -l` | (decoder) | Improvement? |
| **Original (NONE)** | N/A | N/A | 5241912 (5.2M) | 2656624 (2.7M) |  |  |
| snap = "0.2.5" |  | Snappy |  | 2197872 (2.2M) | snap = "0.2.5" | :white_check_mark: |
| flate2 = "1.0.11" | 6 (default) | DEFLATE | 4681344 | 2095472 (2.1M) | inflate = "0.4.5" | :heavy_check_mark: |
|  | 9 (max) | DEFLATE | 4681344 | 2095472 (2.1M) |  | (nothing over level=6) |
| zopfli = "0.4.0" | iter=15 (default) | DEFLATE | 4677240 | 2091376 (2.1M) | inflate = "0.4.5" | compile time ~20x? |
| flate2 = "1.0.11" | 6 (default) | gz | 4721968 | 2128240 (2.1M) | flate2 = "1.0.11" |  |
| flate2 = "1.0.11" | 6 (default) | zlib | 4707816 | 2115952 (2.1M) | flate2 = "1.0.11" |  |
| gzip on unix, else flate2  | default  | gz  | 4717872  | 2124144 (2.1M) |  |  |


Possible compression crates:

* https://crates.io/crates/libflate
* https://crates.io/crates/deflate
* https://crates.io/crates/miniz_oxide
* https://crates.io/crates/zopfli (encoding only; possibly to slow for compiling?)
* https://crates.io/crates/inflate (only decoding: would need an encoder for the build.rs; is only 18kb)
* https://crates.io/crates/zstd

**before:**
```
cargo bloat --release
Compiling ...
Analyzing target/release/sic

 File  .text    Size     Crate Name
 0.6%   2.0% 31.7KiB   inflate inflate::InflateStream::next_state
 0.6%   1.9% 29.5KiB     image image::webp::vp8::VP8Decoder<R>::decode_frame
 0.5%   1.7% 27.0KiB      clap clap::app::parser::Parser::get_matches_with
 0.4%   1.3% 20.5KiB     image jpeg_decoder::decoder::Decoder<R>::decode_internal
 0.4%   1.3% 20.5KiB     image jpeg_decoder::decoder::Decoder<R>::decode_internal
 0.3%   0.9% 14.3KiB      clap clap::app::help::Help::write_arg
 0.3%   0.9% 14.3KiB       png png::decoder::stream::StreamingDecoder::update
 0.3%   0.9% 13.4KiB      clap clap::app::usage::get_required_usage_from
 0.3%   0.9% 13.4KiB      clap clap::app::validator::Validator::validate_matched_args
 0.2%   0.8% 12.2KiB     image image::dynimage::open_impl
 0.2%   0.7% 11.7KiB      clap clap::app::help::Help::_write_parser_help
 0.2%   0.7% 11.5KiB     image tiff::decoder::Decoder<R>::new
 0.2%   0.7% 11.3KiB     image tiff::decoder::ifd::Entry::val
 0.2%   0.7% 10.5KiB     image tiff::decoder::ifd::Entry::val
 0.2%   0.6% 10.1KiB      clap clap::app::validator::Validator::validate
 0.2%   0.6%  9.5KiB      clap clap::app::usage::create_help_usage
 0.2%   0.6%  9.0KiB     image image::dynimage::load_from_memory
 0.2%   0.6%  8.8KiB [Unknown] read_line_info
 0.2%   0.5%  8.4KiB   sic_lib sic_lib::app::cli::cli
 0.1%   0.5%  7.6KiB     image image::webp::vp8::predict_4x4
25.1%  81.3%  1.3MiB           And 2945 smaller methods. Use -n N to show more.
30.9% 100.0%  1.5MiB           .text section size, the file size is 5.0MiB
```
stripped: `2.6M (2656624) target/release/sic`

**after (Snappy):**
```
cargo bloat --release
Compiling ...
Analyzing target/release/sic

 File  .text    Size     Crate Name
 0.7%   2.0% 31.7KiB   inflate inflate::InflateStream::next_state
 0.6%   1.8% 29.5KiB     image image::webp::vp8::VP8Decoder<R>::decode_frame
 0.6%   1.7% 27.0KiB      clap clap::app::parser::Parser::get_matches_with
 0.4%   1.3% 20.5KiB     image jpeg_decoder::decoder::Decoder<R>::decode_internal
 0.4%   1.3% 20.5KiB     image jpeg_decoder::decoder::Decoder<R>::decode_internal
 0.3%   0.9% 14.3KiB      clap clap::app::help::Help::write_arg
 0.3%   0.9% 14.3KiB       png png::decoder::stream::StreamingDecoder::update
 0.3%   0.8% 13.4KiB      clap clap::app::usage::get_required_usage_from
 0.3%   0.8% 13.4KiB      clap clap::app::validator::Validator::validate_matched_args
 0.3%   0.8% 12.2KiB     image image::dynimage::open_impl
 0.2%   0.7% 11.7KiB      clap clap::app::help::Help::_write_parser_help
 0.2%   0.7% 11.5KiB     image tiff::decoder::Decoder<R>::new
 0.2%   0.7% 11.3KiB     image tiff::decoder::ifd::Entry::val
 0.2%   0.7% 10.5KiB     image tiff::decoder::ifd::Entry::val
 0.2%   0.6% 10.1KiB      clap clap::app::validator::Validator::validate
 0.2%   0.6%  9.5KiB      clap clap::app::usage::create_help_usage
 0.2%   0.6%  9.0KiB     image image::dynimage::load_from_memory
 0.2%   0.5%  8.8KiB [Unknown] read_line_info
 0.2%   0.5%  8.4KiB   sic_lib sic_lib::app::cli::cli
 0.2%   0.5%  7.6KiB     image image::webp::vp8::predict_4x4
27.8%  81.5%  1.3MiB           And 2997 smaller methods. Use -n N to show more.
34.1% 100.0%  1.6MiB           .text section size, the file size is 4.6MiB

```

stripped: `2.1M (2197872) target/release/sic`

**After (deflate2, DEFLATE)**:

```
 cargo bloat --release
Compiling ...
Analyzing target/release/sic

 File  .text    Size     Crate Name
 0.7%   2.0% 31.7KiB   inflate inflate::InflateStream::next_state
 0.6%   1.9% 29.5KiB     image image::webp::vp8::VP8Decoder<R>::decode_frame
 0.6%   1.7% 27.0KiB      clap clap::app::parser::Parser::get_matches_with
 0.4%   1.3% 20.5KiB     image jpeg_decoder::decoder::Decoder<R>::decode_internal
 0.4%   1.3% 20.5KiB     image jpeg_decoder::decoder::Decoder<R>::decode_internal
 0.3%   0.9% 14.3KiB      clap clap::app::help::Help::write_arg
 0.3%   0.9% 14.3KiB       png png::decoder::stream::StreamingDecoder::update
 0.3%   0.8% 13.4KiB      clap clap::app::usage::get_required_usage_from
 0.3%   0.8% 13.4KiB      clap clap::app::validator::Validator::validate_matched_args
 0.3%   0.8% 12.2KiB     image image::dynimage::open_impl
 0.3%   0.7% 11.7KiB      clap clap::app::help::Help::_write_parser_help
 0.3%   0.7% 11.5KiB     image tiff::decoder::Decoder<R>::new
 0.2%   0.7% 11.3KiB     image tiff::decoder::ifd::Entry::val
 0.2%   0.7% 10.5KiB     image tiff::decoder::ifd::Entry::val
 0.2%   0.6% 10.1KiB      clap clap::app::validator::Validator::validate
 0.2%   0.6%  9.5KiB      clap clap::app::usage::create_help_usage
 0.2%   0.6%  9.0KiB     image image::dynimage::load_from_memory
 0.2%   0.6%  8.8KiB [Unknown] read_line_info
 0.2%   0.5%  7.6KiB     image image::webp::vp8::predict_4x4
 0.2%   0.5%  7.5KiB     image tiff::decoder::Decoder<R>::new
28.2%  81.4%  1.3MiB           And 2949 smaller methods. Use -n N to show more.
34.6% 100.0%  1.5MiB           .text section size, the file size is 4.5MiB
```

**after (deflate2, gzip)**:
```
 cargo bloat --release   
Compiling ...
Analyzing target/release/sic

 File  .text    Size     Crate Name
 0.7%   2.0% 31.7KiB   inflate inflate::InflateStream::next_state
 0.6%   1.8% 29.5KiB     image image::webp::vp8::VP8Decoder<R>::decode_frame
 0.6%   1.7% 27.0KiB      clap clap::app::parser::Parser::get_matches_with
 0.4%   1.3% 20.5KiB     image jpeg_decoder::decoder::Decoder<R>::decode_internal
 0.4%   1.3% 20.5KiB     image jpeg_decoder::decoder::Decoder<R>::decode_internal
 0.3%   0.9% 14.3KiB      clap clap::app::help::Help::write_arg
 0.3%   0.9% 14.3KiB       png png::decoder::stream::StreamingDecoder::update
 0.3%   0.8% 13.4KiB      clap clap::app::usage::get_required_usage_from
 0.3%   0.8% 13.4KiB      clap clap::app::validator::Validator::validate_matched_args
 0.3%   0.8% 12.2KiB     image image::dynimage::open_impl
 0.3%   0.7% 11.7KiB      clap clap::app::help::Help::_write_parser_help
 0.2%   0.7% 11.5KiB     image tiff::decoder::Decoder<R>::new
 0.2%   0.7% 11.3KiB     image tiff::decoder::ifd::Entry::val
 0.2%   0.7% 10.8KiB miniz_sys tinfl_decompress
 0.2%   0.7% 10.5KiB     image tiff::decoder::ifd::Entry::val
 0.2%   0.6% 10.1KiB      clap clap::app::validator::Validator::validate
 0.2%   0.6%  9.5KiB      clap clap::app::usage::create_help_usage
 0.2%   0.6%  9.0KiB     image image::dynimage::load_from_memory
 0.2%   0.5%  8.8KiB [Unknown] read_line_info
 0.2%   0.5%  7.6KiB     image image::webp::vp8::predict_4x4
28.3%  81.5%  1.3MiB           And 3004 smaller methods. Use -n N to show more.
34.8% 100.0%  1.6MiB           .text section size, the file size is 4.5MiB

```